api/requirements.txt
transformers
torch
sentence-transformers
langchain
faiss-cpu
chromadb
pandas
numpy
tqdm
python-dotenv
fastapi
uvicorn